refactor: rework
