+++ lib/client.js
@@ -11,7 +11,7 @@
 const { replace } = require('./client-hmr');
 const { error, info, refresh, warn } = require('./client-log');
 const { init: initProgress } = require('./overlays/progress');
-const { init: initMinimalProgress } = require('./overlays/minimal-progress');
+const { init: initMinimalProgress } = require('./overlays/progress-minimal');
 const { init: initStatus } = require('./overlays/status');
 
 // eslint-disable-next-line no-undef, no-unused-vars
