+++ lib/overlays/progress-minimal.js
@@ -8,18 +8,16 @@
   The above copyright notice and this permission notice shall be
   included in all copies or substantial portions of this Source Code Form.
 */
-const { addJSText, addCss, addHtml } = require('./util');
+const { addCss, addHtml } = require('./util');
 
-const ns = 'wps_simple_progress';
-
-const progressHtml = `
-<div class="${ns} ${ns}-hidden">
-  <div class="${ns}__bar"></div>
+const ns = 'wps-progress-minimal';
+const html = `
+<div id="${ns}" class="${ns}-hidden">
+  <div class="${ns}-bar"></div>
 </div>
 `;
-
-const progressCss = `
-.${ns} {
+const css = `
+#${ns} {
   position: absolute;
   top: 0;
   left: 0;
@@ -27,39 +25,32 @@ const progressCss = `
   width: 100vw;
 }
 
-.${ns}__bar {
+#${ns}-bar {
   width: 0%;
   height: 4px;
   background-color: rgb(186, 223, 172);
   transition: width 1s ease-in-out;
 }
+
 .${ns}-hidden{
   display: none;
 }
 `;
 
-const progressJS = `
-var elem = document.querySelector(".${ns}__bar");
-var progressWrapper = document.querySelector(".${ns}");
+const update = (percent) => {
+  const bar = document.querySelector(`#${ns}-bar`);
+  bar.style.width = `${percent}%`;
+};
 
-function ${ns}_reset() {
-  progressWrapper.classList.add("${ns}-hidden");
-  elem.style.width = '0%';
-}
-
-function ${ns}_setPercentage(percentage){
-  elem.style.width = percentage + '%';
-}`;
-
-const update = (progress) => {
-  wps_simple_progress_setPercentage(progress);
+const reset = (wrapper) => {
+  wrapper.classList.add(`${ns}-hidden`);
+  setTimeout(() => update(0), 1e3);
 };
 
 const init = (options, socket) => {
   document.addEventListener('DOMContentLoaded', () => {
-    addHtml(progressHtml);
-    addCss(progressCss);
-    addJSText(progressJS);
+    addHtml(html);
+    addCss(css);
   });
 
   socket.addEventListener('message', (message) => {
@@ -70,11 +61,12 @@ const init = (options, socket) => {
     }
 
     const percent = Math.floor(data.percent * 100);
+    const wrapper = document.querySelector(`#${ns}`);
 
-    document.querySelector(`.${ns}`).classList.remove(`${ns}-hidden`);
+    wrapper.classList.remove(`${ns}-hidden`);
 
     if (data.percent === 1) {
-      setTimeout(() => wps_simple_progress_reset(), 5e3);
+      setTimeout(() => reset(wrapper), 5e3);
     }
 
     update(percent);
