+++ lib/overlays/util.js
@@ -21,10 +21,6 @@ const addHtml = (html, parent) => {
   return nodes;
 };
 
-const appendElementHead = (element) => {
-  document.head.appendChild(element);
-};
-
 const addCss = (css) => {
   const style = document.createElement('style');
 
@@ -37,37 +33,7 @@ const addCss = (css) => {
   }
 
   // append the stylesheet for the svg
-  appendElementHead(style);
-};
-
-const createScript = () => {
-  const script = document.createElement('script');
-  script.type = 'text/javascript';
-  return script;
-};
-
-/**
- * Adds src to script tag
- * and adds it into the dom.
- * @param {String} path
- * <script type="text/javascript" src={path}></script>
- */
-const addJSSrc = (path) => {
-  const script = createScript();
-  script.src = path;
-  appendElementHead(script);
-};
-
-/**
- * Adds content inside script tag
- * and adds it into the dom.
- * @param {String} content
- * <script type="text/javascript">content</script>
- */
-const addJSText = (content) => {
-  const script = createScript();
-  script.text = content;
-  document.head.appendChild(script);
+  document.head.appendChild(style);
 };
 
 const socketMessage = (socket, handler) => {
@@ -77,10 +43,4 @@ const socketMessage = (socket, handler) => {
   });
 };
 
-module.exports = {
-  addCss,
-  addHtml,
-  socketMessage,
-  addJSSrc,
-  addJSText
-};
+module.exports = { addCss, addHtml, socketMessage };
